Refactor tabs in event home to their own page
